Move check_predictor() to util; re-order parameterization
